Added link to hb project site, compartmentalized some components out of App.js.
