+++ src/App.css
@@ -14,58 +14,4 @@
   color: rgb(79, 60, 91);
 }
 
-.home-div {
-  background-color: rgba(79, 60, 91, 0.5);
-  width: 100vw;
-  height: 100vh;
-  padding: 25px;
-}
-
-.main-title {
-  
-}
-
-.intro-blurb {
-
-}
-
-.nav-menu {
-
-}
-
-.nav {
-
-}
-
-.nav-list #nav {
-
-}
-
-#contact .nav-item {
-
-}
-
-
-#nav-bar .nav-item {
-  display: inline-block;
-  height: 20vw;
-  width: 20vw;
-  line-height: 20vw;
-  margin: 10px;
-  border: solid 1px;
-  border-color: rgba(79, 60, 91, 1.0);
-  border-radius: 50%;
-  text-align: center;
-  vertical-align: middle;
-}
-
-.nav-link {
-  display: inline;
-  color: rgb(139, 59, 116);
-}
-
-
-#footer-contact-info {
-
-}
