+++ src/components/home.css
@@ -1,0 +1,47 @@
+.home-div {
+  height: 100vh;
+  background-color: rgba(79, 60, 91, 0.5);
+  margin: auto;
+  padding: 15vh;
+}
+
+.main-title {
+  padding-top: 4vh;
+  padding-bottom: 2vh;
+  text-align: center;
+  color: rgba(79, 60, 91, 1.0);
+}
+
+.intro-blurb {
+  text-align: center;
+  color: rgba(79, 60, 91, 1.0);
+}
+
+.nav-menu {
+
+}
+
+.nav {
+
+}
+
+
+#nav-bar .nav-item {
+  display: inline-block;
+  height: 20vw;
+  width: 20vw;
+  line-height: 20vw;
+  margin: 5vh;
+  border: solid 5px;
+  border-color: rgba(79, 60, 91, 1.0);
+  border-radius: 50%;
+  text-align: center;
+  vertical-align: middle;
+  font-size: 2vw;
+}
+
+.nav-link {
+  display: inline;
+  color: rgb(139, 59, 116);
+}
+
