+++ src/components/projects.js
@@ -4,11 +4,16 @@ export default class ProjectContainer extends Component {
   render() {
     return (
       <div className="project-page">
-        <div className='feelin-presidential' >
-          d3 to come!
-        </div>
         <div className="discover-bloom">
-          <a href="/discover-bloom">Bloom, online indigenous language learning.</a>
+          <a href="https://app.discoverbloom.com/">
+            Bloom, online indigenous language learning.
+          </a>
+        </div>
+        <div className='feelin-presidential' >
+          <a href="https://feeling-prez.herokuapp.com/">
+            Feeling Presidential, a sentiment analysis project with
+            Python's Natural Language Toolkit (NLTK), Flask & d3.js
+          </a>
         </div>
       </div>
     )
