+++ src/index.css
@@ -1,9 +1,17 @@
 html, body, root {
-  margin: 0;
-  padding: 0;
   height: 100%;
-  min-height: 100%;
-  max-height: 100%;
   width: 100%;
-  font-family: 'Libre Franklin', sans-serif;;
+  margin: 0;
+  font-family: 'Libre Franklin', sans-serif;
+  background-color: rgba(79, 60, 91, 0.5);
 }
+
+
+/* height: 100%;
+margin: 0;
+padding: 0;
+height: 100%;
+min-height: 100%;
+max-height: 100%;
+width: 100%;
+min-width: 100%; */
